requirements.txt
APScheduler==3.6.3
apsw==3.36.0.post1
argcomplete==1.12.3
attrs==21.2.0
autopep8==1.6.0
bcrypt==3.2.0
Beaker==1.10.0
beautifulsoup4==4.9.3
blivet==3.4.2
blivet-gui==2.3.0
Brlapi==0.8.2
cachetools==4.2.2
cchardet==2.1.7
certifi==2021.10.8
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
chrome-gnome-shell==0.0.0
click==8.0.1
cryptography==3.4.7
css-parser==1.0.6
cupshelpers==1.0
dasbus==1.6
dbus-python==1.2.18
decorator==5.0.9
defusedxml==0.7.1
distro==1.6.0
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
evdev==1.1.2
fedora-third-party==0.8
feedparser==6.0.6
fros==1.1
gpg==1.15.1
greenlet==1.1.2
gssapi==1.6.14
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
humanize==0.5.1
idna==3.2
ifaddr==0.1.7
jeepney==0.7.1
jsonschema==3.2.0
koji==1.27.0
langtable==0.0.56
libcomps==0.1.18
lutris==0.5.9.1
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.6
MarkupSafe==2.0.0
mechanize==0.4.7
msgpack==1.0.3
netifaces==0.10.6
nftables==0.1
odfpy==1.4.1
olefile==0.46
packaging==21.0
paramiko==2.8.1
Paste==3.5.0
pexpect==4.8.0
pid==2.2.3
Pillow==8.3.2
ply==3.11
productmd==1.33
progressbar2==3.53.2
psutil==5.8.0
ptyprocess==0.6.0
pwquality==1.4.4
pyaes==1.6.1
pyasn1==0.4.8
pycairo==1.20.1
pychm==0.8.6
pycodestyle==2.8.0
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.1
pycurl==7.44.1
pyenchant==3.2.2
pygit2==1.6.1
Pygments==2.9.0
PyGObject==3.42.0
pykickstart==3.34
PyNaCl==1.4.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.11.7
PyQt5==5.15.6
PyQt5-sip==12.9.0
PyQtWebEngine==5.15.2
pyrsistent==0.18.0
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.1
python-dotenv==0.19.2
python-magic==0.4.22
python-meh==0.50
python-telegram-bot==13.9
python-utils==2.5.6
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyudev==0.22.0
pyxdg==0.27
PyYAML==5.4.1
regex==2021.10.23
requests==2.26.0
requests-file==1.5.1
requests-ftp==0.3.1
requests-gssapi==1.2.3
rpm==4.17.0
rpmautospec==0.2.5
rsa==4.8
selinux==3.3
sepolicy==3.3
setools==4.4.0
sgmllib3k==1.0.0
simpleaudio==1.0.4
simpleline==1.8.2
six==1.16.0
sos==4.2
soupsieve==2.3.1
SQLAlchemy==1.4.29
systemd-python==234
Telethon==1.24.0
Tempita==0.5.2
texttable==1.6.4
toml==0.10.2
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
webencodings==0.5.1
websocket-client==0.59.0
zeroconf==0.36.9
